pyproject.toml
[tool.poetry]
name = "Programming Exercise"
version = "0.1.0"
description = ""
authors = ["Sana Farooqui <sana.farooqui@ajah.ca>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
replit = "^3.2.4"
pytest = "7.3.2"
pandas = "1.3.4"
PyYAML = "^6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"